// backend/src/types/enums.ts
export enum Role {
  User = 'user',
  Admin = 'admin',
}

export enum City {
  Kyiv = 'Київ',
  Lviv = 'Львів',
  Kharkiv = 'Харків',
  Odesa = 'Одеса',
  Dnipro = 'Дніпро',
  Zaporizhzhia = 'Запоріжжя',
  Vinnytsia = 'Вінниця',
  Chernivtsi = 'Чернівці',
  Ivano_Frankivsk = 'Івано-Франківськ',
  Ternopil = 'Тернопіль',
  Lutsk = 'Луцьк',
  Rivne = 'Рівне',
  Zhytomyr = 'Житомир',
  Khmelnytskyi = 'Хмельницький',
  Cherkasy = 'Черкаси',
  Poltava = 'Полтава',
  Sumy = 'Суми',
  Chernihiv = 'Чернігів',
  Mykolaiv = 'Миколаїв',
  Kherson = 'Херсон',
  Kropyvnytskyi = 'Кропивницький',
  Uzhhorod = 'Ужгород',
}